Move player appropriate scripts to player folder
