--- Assets/Scripts/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b70cec0367616494bb52e19eda6047a2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 200
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
