--- Assets/Scripts/PlayerMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c55b872060617f4459a1e5f4d9f4eb15
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
