--- Assets/Scripts/Posession.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 86ac85168aec4470f92583ad9785c55a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
