--- Assets/Scripts/ProjectileLauncher_1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 134cc88f64e05ae4eb068eb4b30faa94
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
